+++ trunk/simulations-java/simulations/faraday/src/edu/colorado/phet/faraday/control/GraphicSlider.java
@@ -222,7 +222,7 @@ public class GraphicSlider extends GraphicLayerSet {
         if ( value != _value ) {
             
             // Silently clamp the value to the allowed range.
-            _value = (int) MathUtil.clamp( _minimum, value, _maximum );
+            _value = MathUtil.clamp( _minimum, value, _maximum );
 
             // Fire a ChangeEvent to notify listeners that the value has changed.
             fireChangeEvent( new ChangeEvent( this ) );
